Some more steps towards getting Vundle working.
Updated the files that will help Vundle function, as it requires a git
and curl command. GitHub for Windows does not add anything to the system
$PATH, so Vundle, when run from a local Vim install, will not be able to
call git or curl. I'm working on scripts I can add, that will auto-find,
on execution, the latest build of Git installed by GitHub.